# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cakra Prima",
        "description": "Dokumentasi API Kas & Keuangan Cakra Prima",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/installments": {
            "get": {
                "tags": [
                    "Installments"
                ],
                "summary": "Daftar Angsuran",
                "operationId": "d4a30624748baa7766861b1adcf284b3",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Installments"
                ],
                "summary": "Tambah Angsuran",
                "operationId": "3b9440ec0aa3cdc57165717eb699ae3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "due_date",
                                    "amount",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "BPJS"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-31"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 150000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "belum dibayar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/installments/{id}": {
            "get": {
                "tags": [
                    "Installments"
                ],
                "summary": "Detail Angsuran",
                "operationId": "059a3086c06cdc9837ea0c82516347b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Installments"
                ],
                "summary": "Update Angsuran",
                "operationId": "7057368f4c5abd2d496a7df30cce2d71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Installments"
                ],
                "summary": "Hapus Angsuran",
                "operationId": "640c04408805f01ce6ac95c9f220822d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
        "/api/cash-flow-periods/{period}/items": {
            "post": {
                "tags": [
                    "Cash Flow Items"
                ],
                "summary": "Tambah Item Arus Kas Operasional",
                "operationId": "776bca55450111ac6fdc5be75cc81f7c",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "section",
                                    "label",
                                    "amount"
                                ],
                                "properties": {
                                    "section": {
                                        "type": "string",
                                        "example": "modal_awal"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "A"
                                    },
                                    "label": {
                                        "type": "string",
                                        "example": "Modal Awal"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 10000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/cash-flow-periods/{period}/items/{item}": {
            "put": {
                "tags": [
                    "Cash Flow Items"
                ],
                "summary": "Update Item Arus Kas Operasional",
                "operationId": "51c7ecd64bd473f8bf933615784d4565",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "section": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "label": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cash Flow Items"
                ],
                "summary": "Hapus Item Arus Kas Operasional",
                "operationId": "1459eabeaf3f7709a91bf12ab3bfdf8d",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}